requirements.txt
attrs==23.2.0
Automat==22.10.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.5
cssselect==1.2.0
cycler==0.12.1
et-xmlfile==1.1.0
filelock==3.13.4
Flask==3.0.3
fonttools==4.51.0
future==1.0.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.2.0
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
kiwisolver==1.4.5
lxml==5.2.1
MarkupSafe==2.1.5
matplotlib==3.8.4
mysql-connector==2.2.9
mysql-connector-python==8.3.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
parsel==1.9.1
pillow==10.3.0
Protego==0.3.1
pure-sasl==0.6.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
PyHive==0.7.0
PyMySQL==1.1.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
queuelib==1.6.2
requests==2.31.0
requests-file==2.0.0
sasl==0.3.1
scikit-learn==1.4.2
scipy==1.13.0
Scrapy==2.11.1
service-identity==24.1.0
six==1.16.0
soupsieve==2.5
threadpoolctl==3.5.0
thrift==0.20.0
thrift-sasl==0.4.3
tldextract==5.1.2
Twisted==24.3.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
w3lib==2.1.2
Werkzeug==3.0.2
zope.interface==6.3
